requirements.txt
Pillow==3.1.1
argparse==1.2.1
future==0.15.2
python-telegram-bot==3.3
requests==2.9.1
wsgiref==0.1.2
